refactor: modularize quickstart helpers
